+++ filters/olefilters/lib/klaola.h
@@ -33,25 +33,25 @@
 // OLE-Info for the corresponding stream.
 // Normally you won't use it.
 struct OLEInfo {
-    long handle;        // PPS entry number
+    int handle;        // PPS entry number
     QString name;       // Name of the stream
     short nameSize;     // Size of the name
     char type;          // Type of pps
-    long prev;          // Last pps
-    long next;          // Next pps
-    long dir;           // Dir pps
-    long ts1s;          // Timestamp 1, seconds
-    long ts1d;          // Timestamp 1, days
-    long ts2s;          // Timestamp 2, seconds
-    long ts2d;          // Timestamp 2, days
-    long sb;            // Starting block
-    long size;          // Size of property
+    int prev;          // Last pps
+    int next;          // Next pps
+    int dir;           // Dir pps
+    int ts1s;          // Timestamp 1, seconds
+    int ts1d;          // Timestamp 1, days
+    int ts2s;          // Timestamp 2, seconds
+    int ts2d;          // Timestamp 2, days
+    int sb;            // Starting block
+    int size;          // Size of property
 };
 
 // A little bit of OLE-Information. If you want to navigate through
 // the "filesystem" you have to use this struct.
 struct OLENode {
-    long handle;
+    int handle;
     QString name;
     char type;       // 1=Dir, 2=File/Stream, 5=Root Entry
     bool deadDir;    // true, if the dir is a "dead end"
@@ -70,15 +70,15 @@ public:
     const QList<OLENode> parseCurrentDir();
 
     // Wade through the "file system"
-    const bool enterDir(const long &handle);
+    const bool enterDir(const int &handle);
     const bool leaveDir();
-    const QArray<long> currentPath() const;
+    const QArray<int> currentPath() const;
 
-    const OLEInfo streamInfo(const long &handle);
-    const myFile stream(const long &handle);    // Note: data - 512 byte blocks, but
+    const OLEInfo streamInfo(const int &handle);
+    const myFile stream(const int &handle);    // Note: data - 512 byte blocks, but
                                                 // length is set correctly :)
                                                 // Make sure that you delete [] the data!
-    const QArray<long> find(const QString &name, const bool onlyCurrentDir=false);
+    const QArray<int> find(const QString &name, const bool onlyCurrentDir=false);
 
     void testIt();                     // dump some info (similar to "lls"
                                        // of the LAOLA-project)
@@ -88,7 +88,7 @@ private:
     const KLaola &operator=(const KLaola &);
 
     struct OLETree {
-        long handle;
+        int handle;
         short subtree;
     };
 
@@ -97,19 +97,19 @@ private:
     void readSmallBlockDepot();
     void readSmallBlockFile();
     void readRootList();
-    void readPPSEntry(const long &pos, const long &handle);
-    void createTree(const long &handle, const short &index);
-    const unsigned char *readBBStream(const long &start, const bool setmaxSblock=false);
-    const unsigned char *readSBStream(const long &start);
+    void readPPSEntry(const int &pos, const int &handle);
+    void createTree(const int &handle, const short &index);
+    const unsigned char *readBBStream(const int &start, const bool setmaxSblock=false);
+    const unsigned char *readSBStream(const int &start);
 
-    inline const long nextBigBlock(const long &pos);
-    inline const long nextSmallBlock(const long &pos);
-    inline const unsigned short read16(const long &i);
-    inline const unsigned long read32(const long &i);
+    inline const int nextBigBlock(const int &pos);
+    inline const int nextSmallBlock(const int &pos);
+    inline const unsigned short read16(const int &i);
+    inline const unsigned int read32(const int &i);
 
     QList<QList<OLETree> > treeList;
     QList<OLEInfo> ppsList;
-    QArray<long> path;
+    QArray<int> path;
 
     bool ok;        // is the file OK?
 
@@ -118,11 +118,11 @@ private:
     unsigned char *smallBlockDepot;
     unsigned char *smallBlockFile;
 
-    unsigned long maxblock;          // maximum number of big-blocks
-    unsigned long maxSblock;         //                   small-blocks
-    unsigned long num_of_bbd_blocks; // number of big block depot blocks
-    unsigned long root_startblock;   // Root chain's first big block
-    unsigned long sbd_startblock;    // small block depot's first big block
-    unsigned long *bbd_list;  //array of num_of_bbd_blocks big block numbers
+    unsigned int maxblock;          // maximum number of big-blocks
+    unsigned int maxSblock;         //                   small-blocks
+    unsigned int num_of_bbd_blocks; // number of big block depot blocks
+    unsigned int root_startblock;   // Root chain's first big block
+    unsigned int sbd_startblock;    // small block depot's first big block
+    unsigned int *bbd_list;  //array of num_of_bbd_blocks big block numbers
 };
 #endif // KLAOLA_H
