+++ filters/olefilters/winword97/asumy.h
@@ -23,7 +23,7 @@
 #define ASUMY_H
 
 struct ASUMY {
-    long lLevel;
+    int lLevel;
 
     bool ok;
 };
