+++ filters/olefilters/winword97/bte.h
@@ -23,8 +23,8 @@
 #define BTE_H
 
 struct BTE {
-    long pn:22;
-    long unused:10;
+    int pn:22;
+    int unused:10;
 
     bool ok;
 };
