+++ filters/olefilters/winword97/misc.cc
@@ -23,7 +23,7 @@ const unsigned short read16(const unsigned char *d) {
     return ( (*(d+1) << 8) + *d );
 }
 
-const unsigned long read32(const unsigned char *d) {
+const unsigned int read32(const unsigned char *d) {
     return ( (read16(d+2) << 16) + read16(d) );
 }
 
@@ -34,7 +34,7 @@ const short char2uni(const unsigned char &c) {
         return CP2UNI[c-0x80];
 }
 
-void align2(unsigned long &adr) {
+void align2(unsigned int &adr) {
     if((adr%2)!=0)
         ++adr;
 }
